Fix issue with loback closing down in devtools
Still not sure what causes it, but with devtools the logging system
seems to shutdown before the app has stopped, so we miss the
output from the listener we installed and can't measure the
startup time.

Also upgrades to snapshots by default, so in the IDE you can
test the new features.